Simple Sprite. ATI fix 2.

+++ assets/Shaders/SimpleSprite/SimpleSprite.vert
@@ -25,8 +25,8 @@ void main(){
    gl_Position = g_WorldViewProjectionMatrix * vec4(inPosition, 1.0);
    texCoordAni = inTexCoord;
 
-	int iNumTilesU = float(m_numTilesU);
-	int iNumTilesV = float(m_numTilesV);
+	int iNumTilesU = int(m_numTilesU);
+	int iNumTilesV = int(m_numTilesV);
 
 	int numTilesTotal = iNumTilesU * iNumTilesV;
 	int selectedTile = 1.0;
@@ -35,7 +35,7 @@ void main(){
 selectedTile += g_Time*m_Speed;
 
 	// the "1 - " bit is because otherwise it goes from right to left
-	texCoordAni.x = -(1 - ((texCoordAni.x + mod(int(selectedTile),  int(iNumTilesU))) / iNumTilesU)); ///selectedTile;
+	texCoordAni.x = -(1.0 - ((texCoordAni.x + mod(int(selectedTile),  int(iNumTilesU))) / iNumTilesU)); ///selectedTile;
         texCoordAni.y = ((-texCoordAni.y - int(selectedTile / iNumTilesU)) / iNumTilesV); ///selectedTile;
